밸런싱 완료
내가 PlayerStatsManager에 써놓은 게 가장 좋다고 생각하는 수치임.

+++ Assets/Scripts/DroppedItem/DroppedFeatherController.cs
@@ -4,7 +4,7 @@ public class DroppedFeatherController : DroppedItemController
 {
 
     [SerializeField]
-    int addMoveSpeed = 1;
+    float addMoveSpeed = 1;
 
 
     protected override void Update()
